Use jenkins docker agent for all pipelines if configured
