+++ vars/withParameterizedPipeline.groovy
@@ -58,9 +58,13 @@ def call(type, String product, String component, String environment, String subs
   def deploymentTargetList = deploymentTargets.split(',') as List
   AKSSubscriptions aksSubscriptions = new AKSSubscriptions(this)
 
-  node {
-    def slackChannel = new TeamConfig(this).getBuildNoticesSlackChannel(product)
+  def teamConfig = new TeamConfig(this)
+  String agentType = teamConfig.getBuildAgentType(product)
+
+  node(agentType) {
+    def slackChannel = teamConfig.getBuildNoticesSlackChannel(product)
     try {
+      dockerAgentSetup(product)
       env.PATH = "$env.PATH:/usr/local/bin"
 
       stage('Checkout') {
